AVRO-3364 Breakout subclasses for GenericDatumReader
