+++ lang/csharp/src/apache/main/Generic/GenericDatumReader.cs
@@ -15,23 +15,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-using System;
-using System.Collections.Generic;
-using Avro.IO;
 
 namespace Avro.Generic
 {
     /// <summary>
-    /// <see cref="PreresolvingDatumReader{T}"/> for reading data to <see cref="GenericRecord"/>
+    /// <see cref="PreresolvingDatumReader{T}" /> for reading data to <see cref="GenericRecord" />
     /// classes or primitives.
     /// <see cref="PreresolvingDatumReader{T}">For more information about performance considerations
     /// for choosing this implementation</see>.
     /// </summary>
     /// <typeparam name="T">Type to deserialize data into.</typeparam>
-    public class GenericDatumReader<T> : PreresolvingDatumReader<T>
+    /// <seealso cref="PreresolvingDatumReader&lt;T&gt;" />
+    public partial class GenericDatumReader<T> : PreresolvingDatumReader<T>
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="GenericDatumReader{T}"/> class.
+        /// Initializes a new instance of the <see cref="GenericDatumReader{T}" /> class.
         /// </summary>
         /// <param name="writerSchema">Schema that was used to write the data.</param>
         /// <param name="readerSchema">Schema to use to read the data.</param>
@@ -85,143 +83,6 @@ namespace Avro.Generic
         protected override FixedAccess GetFixedAccess(FixedSchema readerSchema)
         {
             return new GenericFixedAccess(readerSchema);
-        }
-
-        class GenericEnumAccess : EnumAccess
-        {
-            private EnumSchema schema;
-
-            public GenericEnumAccess(EnumSchema schema)
-            {
-                this.schema = schema;
-            }
-
-            public object CreateEnum(object reuse, int ordinal)
-            {
-                if (reuse is GenericEnum)
-                {
-                    var ge = (GenericEnum) reuse;
-                    if (ge.Schema.Equals(this.schema))
-                    {
-                        ge.Value = this.schema[ordinal];
-                        return ge;
-                    }
-                }
-                return new GenericEnum(this.schema, this.schema[ordinal]);
-            }
-        }
-
-        internal class GenericRecordAccess : RecordAccess
-        {
-            private RecordSchema schema;
-
-            public GenericRecordAccess(RecordSchema schema)
-            {
-                this.schema = schema;
-            }
-
-            public object CreateRecord(object reuse)
-            {
-                GenericRecord ru = (reuse == null || !(reuse is GenericRecord) || !(reuse as GenericRecord).Schema.Equals(this.schema)) ?
-                    new GenericRecord(this.schema) :
-                    reuse as GenericRecord;
-                return ru;
-            }
-
-            public object GetField(object record, string fieldName, int fieldPos)
-            {
-                object result;
-                if(!((GenericRecord)record).TryGetValue(fieldPos, out result))
-                {
-                    return null;
-                }
-                return result;
-            }
-
-            public void AddField(object record, string fieldName, int fieldPos, object fieldValue)
-            {
-                ((GenericRecord)record).Add(fieldName, fieldValue);
-            }
-        }
-
-        class GenericFixedAccess : FixedAccess
-        {
-            private FixedSchema schema;
-
-            public GenericFixedAccess(FixedSchema schema)
-            {
-                this.schema = schema;
-            }
-
-            public object CreateFixed(object reuse)
-            {
-                return (reuse is GenericFixed && (reuse as GenericFixed).Schema.Equals(this.schema)) ?
-                    reuse : new GenericFixed(this.schema);
-            }
-
-            public byte[] GetFixedBuffer( object f )
-            {
-                return ((GenericFixed)f).Value;
-            }
-        }
-
-        class GenericArrayAccess : ArrayAccess
-        {
-            public object Create(object reuse)
-            {
-                    return (reuse is object[]) ? reuse : new object[0];
-            }
-
-            public void EnsureSize(ref object array, int targetSize)
-            {
-                if (((object[])array).Length < targetSize)
-                    SizeTo(ref array, targetSize);
-            }
-
-            public void Resize(ref object array, int targetSize)
-            {
-                SizeTo(ref array, targetSize);
-            }
-
-            public void AddElements( object arrayObj, int elements, int index, ReadItem itemReader, Decoder decoder, bool reuse )
-            {
-                var array = (object[]) arrayObj;
-                for (int i = index; i < index + elements; i++)
-                {
-                    array[i] = reuse ? itemReader(array[i], decoder) : itemReader(null, decoder);
-                }
-            }
-
-            private static void SizeTo(ref object array, int targetSize)
-            {
-                var o = (object[]) array;
-                Array.Resize(ref o, targetSize);
-                array = o;
-            }
-        }
-
-        class GenericMapAccess : MapAccess
-        {
-            public object Create(object reuse)
-            {
-                if (reuse is IDictionary<string, object>)
-                {
-                    var result = (IDictionary<string, object>)reuse;
-                    result.Clear();
-                    return result;
-                }
-                return new Dictionary<string, object>();
-            }
-
-            public void AddElements(object mapObj, int elements, ReadItem itemReader, Decoder decoder, bool reuse)
-            {
-                var map = (IDictionary<string, object>)mapObj;
-                for (int i = 0; i < elements; i++)
-                {
-                    var key = decoder.ReadString();
-                    map[key] = itemReader(null, decoder);
-                }
-            }
         }
     }
 }
